Reza Shakoori interview

+++ src/include/access/hd_work_mgr.h
@@ -26,7 +26,7 @@ typedef struct sPxfFragmentStatsElem
 {
 	int numFrags;
 	int firstFragSize; /* size of the first fragment */
-	int totalSize; /* size of the total datasource */
+	long long totalSize; /* size of the total datasource */
 } PxfFragmentStatsElem;
 PxfFragmentStatsElem *get_pxf_fragments_statistics(char *uri, Relation rel, StringInfo err_msg);
